Fix GIF resizing issue in !sim (Finally!) Finally :) The GIF resizing issue seems to be fixed.
We calculate the min & max of x, y coordinates for all gens.
We use the min x and y instead of the top-left corner for the pivot.
We also use the (max x - min x) * (max y - min y) as the frame size,
to bound the pattern activity across all generations.

The drawing algorithm is also changed (quite a bit) for the ease of
implementateion with the changed parameters.