Rewrote & greatly improved cogs.role_stuffs
Basically just reworked the thing from the ground up because I realized its actual selling point (users assigning roles/colors to themselves) clashed irrepairably with my original vision (a simple bot to allow admin to assign roles to people); my last commit was an attempt to tack this new realization onto my old code in a the most slapdash manner I could come up with, and it was really just awful because they don't really mix like that so I had a bunch of stupid little bugs/oversights and exception handlers coming out the arse and increasingly-copious amounts of asynchronous sticky-tack to try to by sheer force of will compel it to somehow not be trash... which didn't work. So here's a total rewrite instead. This is also the version I managed to delete by accident (see the function that has the ability to edit __file__? Yeah, I was an idiot and didn't realize first time around that you have to reprint EVERY line with fileinput.input(). rip.) but I very luckily managed to rewrite it by hand and from memory today after uncompyle6 and other decompile utilities proved nigh useless for py3.6.\nOne last thing: I believe I'm going to move this to and perhaps start with some Discord board games on a new bot sometime in the maybe-near future, because creating a bot that edits roles isn't exactly a CA-related endeavor as the rest of Caterer is.